requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.12.7
cloudinary==1.30.0
dj-database-url==1.2.0
Django==4.1.3
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-redis==5.2.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2==2.9.5
python-decouple==3.6
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.13
views==0.3
whitenoise==6.2.0
